modesetting: fix tearfree with DRI2/Present flipping simultaneously
Fix the return codes for queue_flip_on_crtc() and skip tearfree when
DRI2/Present are flipping on the screen to avoid each one stepping on the
others' toes. When there's a tearfree flip pending at the same time
DRI2 or Present want to flip, stall until the tearfree flip is finished.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>